New system for placing latex spreads
From now own youonly have to specify the Id of text, plot or chart only
once for chart_spread -function that will generate the latex output for
the whole regular spread. This will simplify the way to content of each
spread is written and makes reordering the spreads easier, too. Will
work on this tomorrow morning a bit more.