Added sample spreads for Amy
In practise each block of code beginning with <<something, eval=FALSE>> and ending with @ creates a spread with charts adn text on the left and map on the right page.

Thought the underlying data nor the actual plots do not exists in the project yet, these spreads could be defined already for this file. Codes do not matter, but a comment after # with variable, plot type and time frame would be great!